requirements.txt
flask
flask-cors
supabase
python-dotenv
bcrypt
flask-jwt-extended
httpx
